requirements.txt
Django==4.2.1
djangorestframework==3.14.0
dj-rest-auth==4.0.0
dj-rest-auth[with_social]
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.6